Change interface to always return a value.
This gets around returning different things when a resource or method can't be
found. Also, change the name of the method that does the work to C<dispatch>.
If a resource or method can't be found, it returns a PSGI array with the
necessary information. This makes it easier to use now, too. All one has to do
for a Plack app is:

    sub { $router->dispatch($env) };

Of course users will want custom 404 and 405 handlers, so add support for them
with the `missing` keyword (which may be renamed if I think of something
better). This defines a code block at the same level as resources to execute for
404 and 405 responses.

The `match` method is stil around, and returns a hash ref, but I might get rid
of it altogether.

While at it, add the `TRACE` and `CONNECT` methods. No idea what they do, but
they're in HTTP 1.1, so might as well be complete.